requirements.txt
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
Pillow>=10.0.0
ExifRead>=3.0.0